Moved to shared and added to menu #127
