+++ src/main/webapp/app/admin/dashboard/request/request.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'up-request',
+  templateUrl: './request.component.html',
+  styles: []
+})
+export class RequestComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
