+++ src/main/webapp/app/admin/dashboard/request/request.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RequestComponent } from './request.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [RequestComponent]
+})
+export class RequestModule { }
