+++ src/main/webapp/app/entities/request/index.ts
@@ -1,6 +1,6 @@
-export * from './request.model';
+export * from '../../shared/request/request.model';
 export * from './request-popup.service';
-export * from './request.service';
+export * from '../../shared/request/request.service';
 export * from './request-dialog.component';
 export * from './request-delete-dialog.component';
 export * from './request-detail.component';
