+++ src/main/webapp/app/entities/request/request-dialog.component.ts
@@ -6,9 +6,9 @@ import { Observable } from 'rxjs/Observable';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { JhiEventManager, JhiAlertService, JhiDataUtils } from 'ng-jhipster';
 
-import { Request } from './request.model';
+import { Request } from '../../shared/request/request.model';
 import { RequestPopupService } from './request-popup.service';
-import { RequestService } from './request.service';
+import { RequestService } from '../../shared/request/request.service';
 import { Property, PropertyService } from '../property';
 
 @Component({
