+++ src/main/webapp/app/layouts/admin-sidebar/admin-sidebar.component.html
@@ -36,7 +36,12 @@
             </li>
             <li class="nav-item">
                 <a routerLink="/dashboard/real-estates" routerLinkActive="active" [routerLinkActiveOptions]="{ exact: true }" class="nav-link">
-                    <i class="fa fa-building"></i><span>Real Estates</span>
+                    <i class="fa fa-home"></i><span>Real Estates</span>
+                </a>
+            </li>
+            <li class="nav-item">
+                <a routerLink="/dashboard/requests" routerLinkActive="active" [routerLinkActiveOptions]="{ exact: true }" class="nav-link">
+                    <i class="fa fa-bell"></i><span>Requests</span>
                 </a>
             </li>
         </ul>
