+++ src/test/javascript/spec/app/entities/request/request.service.spec.ts
@@ -3,7 +3,7 @@ import { TestBed, getTestBed } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { JhiDateUtils } from 'ng-jhipster';
 
-import { RequestService } from '../../../../../../main/webapp/app/entities/request/request.service';
+import { RequestService } from '../../../../../../main/webapp/app/shared/request/request.service';
 import { SERVER_API_URL } from '../../../../../../main/webapp/app/app.constants';
 
 describe('Service Tests', () => {
